+++ R/cross-val.R
@@ -103,7 +103,6 @@ ll_sdmTMB <- function(object, withheld_y, withheld_mu) {
 #'   object. This vector is appended to `TRUE` and passed to the argument
 #'   `future.globals` in [future.apply::future_lapply()]. Useful if global
 #'   objects are used to specify arguments like priors, families, etc.
-#' @param spde **Depreciated.** Use `mesh` instead.
 #' @param ... All other arguments required to run [sdmTMB()] model with the
 #'   exception of `weights`, which are used to define the folds.
 #'
@@ -154,6 +153,11 @@ ll_sdmTMB <- function(object, withheld_y, withheld_mu) {
 #' - Fit data to time steps 1 to 6, predict and validate step 8.
 #' - Fit data to time steps 1 to 7, predict and validate step 9.
 #'
+#' Note these are time steps as they are presented in order in the data.
+#' For example, in the `pcod` data example below steps between data points
+#' are not always one year but an `lfo_forecast = 2` forecasts 2 time
+#' steps as presented not two years.
+#'
 #' See example below.
 #'
 #' @examples
@@ -175,6 +179,7 @@ ll_sdmTMB <- function(object, withheld_y, withheld_mu) {
 #' m_cv$models[[1]]
 #' m_cv$max_gradients
 #'
+#'
 #' \donttest{
 #' # Create mesh each fold:
 #' m_cv2 <- sdmTMB_cv(
@@ -190,23 +195,24 @@ ll_sdmTMB <- function(object, withheld_y, withheld_mu) {
 #'   family = tweedie(link = "log"),
 #'   fold_ids = rep(seq(1, 3), nrow(pcod))[seq(1, nrow(pcod))]
 #' )
-#
-# # LFOCV:
-# m_lfocv <- sdmTMB_cv(
-#   present ~ s(year, k = 4),
-#   data = pcod,
-#   mesh = mesh,
-#   lfo = TRUE,
-#   lfo_forecast = 2,
-#   lfo_validations = 3,
-#   family = binomial(),
-#   spatiotemporal = "off",
-#   time = "year" # must be specified
-# )
-#
-# # See how the LFOCV folds were assigned:
-# example_data <- m_lfocv$models[[1]]$data
-# table(example_data$cv_fold, example_data$year)
+#'
+#' # LFOCV:
+#' m_lfocv <- sdmTMB_cv(
+#'   present ~ s(year, k = 4),
+#'   data = pcod,
+#'   lfo = TRUE,
+#'   lfo_forecast = 2,
+#'   lfo_validations = 3,
+#'   family = binomial(),
+#'   mesh = mesh,
+#'   spatial = "off", # fast example
+#'   spatiotemporal = "off", # fast example
+#'   time = "year" # must be specified
+#' )
+#'
+#' # See how the LFOCV folds were assigned:
+#' fold_table <- table(m_lfocv$data$cv_fold, m_lfocv$data$year)
+#' fold_table
 #' }
 sdmTMB_cv <- function(
     formula, data, mesh_args, mesh = NULL, time = NULL,
@@ -217,15 +223,10 @@ sdmTMB_cv <- function(
     parallel = TRUE,
     use_initial_fit = FALSE,
     future_globals = NULL,
-    spde = deprecated(),
     ...) {
   if (k_folds < 1) cli_abort("`k_folds` must be >= 1.")
 
-  if (is_present(spde)) {
-    deprecate_warn("0.0.20", "sdmTMB_cv(spde)", "sdmTMB_cv(mesh)")
-  } else {
-    spde <- mesh
-  }
+  spde <- mesh
   data[["_sdm_order_"]] <- seq_len(nrow(data))
   constant_mesh <- missing(mesh_args)
   if (missing(mesh_args)) mesh_args <- NULL
@@ -244,11 +245,10 @@ sdmTMB_cv <- function(
       # Create lfo_validations + 1 folds, ordered sequentially
       data$cv_fold <- 1
       t_validate <- sort(unique(data[[time]]), decreasing = TRUE)
-      for (t in seq(1, lfo_validations)) {
+      for (t in seq(1, lfo_validations+lfo_forecast)) {
         # fold id increasing order + forecast
         data$cv_fold[data[[time]] == t_validate[t]] <- lfo_validations - t + 1 + lfo_forecast
       }
-      data$cv_fold <- as.numeric(as.factor(data$cv_fold))
     } else {
       dd <- lapply(split(data, data[[time]]), function(x) {
         x$cv_fold <- sample(rep(seq(1L, k_folds), nrow(x)), size = nrow(x))