+++ R/tmb-sim.R
@@ -345,6 +345,9 @@ sdmTMB_simulate <- function(formula,
 #'   lets you parse out whatever elements you want from the simulation.
 #'   Not usually needed.
 #' @param observation_error Logical. Simulate observation error?
+#' @param size A vector of size (trials) in the case of a binomial family with
+#'   `newdata` specified. If left `NULL` with `newdata`, will be assumed to
+#'   be a vector of 1s.
 #' @param silent Logical. Silent?
 #' @param ... Extra arguments passed to [predict.sdmTMB()]. E.g., one may wish
 #'   to pass an `offset` argument if `newdata` are supplied in a model with an
@@ -397,6 +400,7 @@ simulate.sdmTMB <- function(object, nsim = 1L, seed = sample.int(1e6, 1L),
                             mcmc_samples = NULL,
                             return_tmb_report = FALSE,
                             observation_error = TRUE,
+                            size = NULL,
                             silent = FALSE,
                             ...) {
   set.seed(seed)
@@ -426,7 +430,7 @@ simulate.sdmTMB <- function(object, nsim = 1L, seed = sample.int(1e6, 1L),
     # generate prediction TMB data list
     p <- predict(object, newdata = newdata, return_tmb_data = TRUE, ...)
     # move data elements over
-    p <- move_proj_to_tmbdat(p, object, newdata)
+    p <- move_proj_to_tmbdat(p, object, newdata, called_by_simulate = TRUE, size = size)
     p$sim_re <- tmb_dat$sim_re
     tmb_dat <- p
   }
@@ -479,14 +483,6 @@ simulate.sdmTMB <- function(object, nsim = 1L, seed = sample.int(1e6, 1L),
 
   if (!return_tmb_report) {
     if (is_delta(object)) {
-      if (object$family[[1]]$family == "binomial" &&
-          object$family$type == "standard" &&
-          !observation_error) {
-        ret <- lapply(ret, function(.x) {
-          .x[,1] <- stats::plogis(.x[,1]) # using robust dbinom in logit space
-          .x
-        })
-      }
       if (is.na(model[[1]])) {
         ret <- lapply(ret, function(.x) .x[,1] * .x[,2])
       } else if (model[[1]] == 1) {
@@ -495,13 +491,6 @@ simulate.sdmTMB <- function(object, nsim = 1L, seed = sample.int(1e6, 1L),
         ret <- lapply(ret, function(.x) .x[,2])
       } else {
         cli_abort("`model` argument isn't valid; should be NA, 1, or 2.")
-      }
-    }
-
-    if (!is_delta(object) && !observation_error) {
-      if (object$family$family == "binomial") {
-        # using robust dbinom in logit space
-        ret <- lapply(ret, function(.x) stats::plogis(.x))
       }
     }